+++ torch_brain/utils/binning.py
@@ -11,7 +11,7 @@ def bin_spikes(
     max_spikes: Optional[int] = None,
     right: bool = True,
     eps: float = 1e-3,
-    dtype: np.dtype = np.int64,
+    dtype: np.dtype = np.int32,
 ) -> np.ndarray:
     r"""Bins spikes into time bins of size `bin_size`. If the total time spanned by
     the spikes is not a multiple of `bin_size`, the spikes are truncated to the nearest
@@ -35,7 +35,7 @@ def bin_spikes(
         right: Decide which side gets truncated when duration is not
             a multiple of ``bin_size``. If ``True``, excess spikes are truncated from the left edge.
         eps: Small numerical margin used during bin assignment.
-        dtype: Data type of the output binned array. (default ``np.int64``)
+        dtype: Data type of the output binned array. (default ``np.int32``)
 
     Returns:
         Binned spike counts with shape ``(T, N)``, where ``T`` is the number of
